+++ lib/Fhaculty/Graph/Vertex.php
@@ -539,23 +539,6 @@ class Vertex extends Layoutable
     }
 
     /**
-     * checks whether this vertex has a loop (edge to itself)
-     *
-     * @return boolean
-     * @uses Edge::isLoop()
-     */
-    public function hasLoop()
-    {
-        foreach ($this->edges as $edge) {
-            if ($edge->isLoop()) {
-                return true;
-            }
-        }
-
-        return false;
-    }
-
-    /**
      * destroy vertex and all edges connected to it and remove reference from graph
      *
      * @uses Edge::destroy()